backend/requirements.txt
bcrypt==4.1.2
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
Jinja2==3.1.3
# jwt==1.3.1
PyJWT==2.8.0
MarkupSafe==2.1.5
mysqlclient==2.2.4
requests==2.31.0
rsa==4.9
SQLAlchemy==2.0.27
Werkzeug==3.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
git+https://github.com/maxcountryman/flask-uploads.git@f66d7dc